Fixed a packaging bug
Accidental refactoring of a function caused issues, which only occured when trying to package the game. Took a while to understand what could the  reason be, but thanks to the callstack I found out the reason. Had to delete a method and create it again. Packaged game worked with Steam. Need to do something with the audio. Footsteps attenuation needs to be fixed.