Warmup state
Added a warmup state where the player is a spectator. While in warmup, a countdown to the match start is visible. Countdown time is synchronized with all players. In the Q&A some people experienced problems with a packaged game, where the synchronization was way off. Might be because of timer starts when lobby level is loaded.