Imports and added an SMG
Imported smg meshes and whatnot, as wells as 2 sound cues. Portal sound and interaction sound. SMG is completed. Created a new animation for it. Added a new physics asset just for the smg strap and added physics to it. Fixed a few bugs and made base accuracy a variable on the gun.