More weapons
Added a sniper rifle and a grenade launcher. Refactored projectile code from rocketprojectile to projectile class, so that the grenade projectile can use the functionality.
Sniper has a scope. Hides player when zooming and uses spread when not zoomed.
Created a new smoke trail effect with niagara for the grenade.