Pickup Spawnpoint
Added a class for pickup spawns. You can select which pickups are spawned by a specific spawn point. When pickup is destroyed a new one will spawn after a specified amount of time. Note: Needed to add a timer to pickup so that destroy wouldn't be called as soon as pickup spawns, so that there would be enough time to bind a function to OnDestroyed