HUD fixes and default weapon
Fixed a bug where sniper scatter wasn't working on a client. Fixed bugs where HUD wasn't properly set for client or the host when spawning. Added a default weapon which spawns when player is spawned, so the player isn't ever without a weapon.
Fixed a nasty crash that happened right after 2nd player joined. The callstack gave a confusing function call as the culprit. Claimed that a nullptr was being accessed even though that shouldn't have been possible. Debugger showed that a value passed if checks even when null. Property was set as uproperty so this shouldn't have happened. Figured that it had something to do with the HUD. I was probably right, as I made a few changes to the HUD and weapon class and the problem disappeared.
Maybe someone wiser might know. *Shrugs*