ChatBox #2
Chat box is good enough for the optional challenge. Works as intended. Took me a few hours of debugging to try to get MultiLine elements to work with the chat box but it looks to me like an engine bug. Tried to make a forum post but posting a question was bugged.